feat: added two ways to found dichotomy result

+++ src/main/java/dev/tssvett/service/DichotomyResult.java
@@ -1,8 +1,8 @@
 package dev.tssvett.service;
 
 public record DichotomyResult(
-    double dichotomyResult,
-    double epsilon,
-    double iterations
+        double dichotomyResult,
+        double epsilon,
+        int iterations
 ) {
 }
